eval: Replace with listsetvar with mklocal/setvareq
This patch replaces listsetvar with mklocal/setvareq.  As we now
determine special built-in status prior to variable assignment, we
no longer have to do a second pass listsetvar.  Instead we will
call setvareq directly instead of mklocal when necessary.

In order to do this mklocal can now take a flag in order to mark
a variable for export.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>